Moved the check on zero losses in a view
