+++ openquake/calculators/event_based.py
@@ -226,12 +226,7 @@ class EventBasedCalculator(base.HazardCalculator):
 
     def set_param(self, **kw):
         oq = self.oqparam
-        # set the minimum_intensity
-        if hasattr(self, 'crmodel') and not oq.minimum_intensity:
-            # infer it from the risk models if not directly set in job.ini
-            oq.minimum_intensity = self.crmodel.min_iml
-        min_iml = oq.min_iml
-        if oq.ground_motion_fields and sum(min_iml.values()) == 0:
+        if oq.ground_motion_fields and oq.min_iml.sum() == 0:
             logging.warning('The GMFs are not filtered: '
                             'you may want to set a minimum_intensity')
         else:
