+++ openquake/calculators/scenario_risk.py
@@ -200,41 +200,3 @@ class EventBasedRiskCalculator(base.RiskCalculator):
                 'Due to rounding errors inherent in floating-point arithmetic,'
                 ' agg_losses != sum(avg_losses): %s != %s\nsee %s',
                 agglosses.mean(), sumlosses.mean(), url)
-        try:
-            if sum(oq.minimum_asset_loss.values()) == 0:
-                # check losses are nonzero where avg_gmf is nonzero
-                self.check_losses(oq)
-        except Exception as exc:
-            logging.error('Could not run the sanity check: %s' % exc,
-                          exc_info=True)
-
-    def check_losses(self, oq):
-        """
-        Sanity check on avg_losses and avg_gmf
-        """
-        avg_gmf = self.datastore['avg_gmf'][0]
-        asset_df = self.datastore.read_df('assetcol/array', 'site_id')
-        for col in asset_df.columns:
-            if not col.startswith('value-'):
-                del asset_df[col]
-        values_df = asset_df.groupby(asset_df.index).sum()
-        avglosses = self.avglosses.sum(axis=1) / self.R  # shape (A, L)
-        dic = dict(site_id=self.assetcol['site_id'])
-        for lti, lname in enumerate(oq.loss_names):
-            dic[lname] = avglosses[:, lti]
-        losses_df = pandas.DataFrame(dic).groupby('site_id').sum()
-        sids = losses_df.index.to_numpy()
-        avg_gmf = avg_gmf[sids]
-        nonzero_gmf = (avg_gmf > 0).any(axis=1)
-        nonzero_losses = (losses_df > 0).to_numpy().any(axis=1)
-        bad, = numpy.where(nonzero_gmf != nonzero_losses)
-        # this happens in scenario_risk/case_shakemap and case_3
-        msg = 'Site #%d is suspicious:\navg_gmf=%s\navg_loss=%s\nvalues=%s'
-        for idx in bad:
-            sid = sids[idx]
-            logging.warning(msg, sid, dict(zip(oq.all_imts(), avg_gmf[sid])),
-                            _get(losses_df, sid), _get(values_df, sid))
-
-
-def _get(df, sid):
-    return df.loc[sid].to_dict()
